Allow MGE to work with 3D mass profiles
Turns out a very simple modification can be make to the MGE code to allow it to take in a 3D radial mass profile rather than a project 2D mass profile.  If the 3D profile is decomposed in the typical way, the 2D project is trivial to compute since the project of each gaussian is just another gaussian with its amplitude scaled by sqrt(2*pi)*sigma.

If the `tree_d` keyword is set in the MGE it will add in this factor and proceed as normal.  This enables the family of NFW profiles to be coded up far easier as they all have easy to write down 3D profiles (and typically non-analytic 2D projections).

To leverage this new ability the NFW_ellipse_kappa profile has been written with far less (and far easier to read) code.  Along side this the truncated NFW, cuspy NFW, and cuspy halo mass models have also be included.  The last two are both called "generalized NFW profiles" depending on the paper, I went with the Keeton 2002 names to try and avoid confusion.